+++ src/compas/geometry/brep/brep.py
@@ -923,11 +923,6 @@ class Brep(Geometry):
         -------
         None
 
-        Raises
-        ------
-        :class:`~compas.geometry.BrepInvalidError`
-            If the operation fails.
-
         """
         raise NotImplementedError
 
@@ -1141,6 +1136,26 @@ class Brep(Geometry):
         Returns
         -------
         tuple[list[:class:`compas.geometry.BrepFace`]]
+
+        """
+        raise NotImplementedError
+
+    def cap_planar_holes(self, tolerance=None):
+        """Cap all planar holes in the Brep.
+
+        Parameters
+        ----------
+        tolerance : float, optional
+            The precision to use for the operation. Defaults to `TOL.absolute`.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        BrepError
+            If the operation fails.
 
         """
         raise NotImplementedError